Move routes to micro app root, use lodash to set errors
