+++ src/apps/investments/controllers/create/create-profile.js
@@ -1,13 +1,10 @@
+const { set } = require('lodash')
+
 const { createCompanyProfile } = require('./../../../companies/apps/investments/large-capital-profile/repos.js')
+
 const createProfile = async (req, res, next) => {
   if (!req.body.investor_company) {
-    res.locals.form = Object.assign({}, res.locals.form, {
-      errors: {
-        'messages': {
-          'investor_company': ['Enter a company name'],
-        },
-      },
-    })
+    set(res.locals, 'errors.messages.investor_company', ['Enter a company name'])
     return next()
   }
 
@@ -15,12 +12,12 @@ const createProfile = async (req, res, next) => {
     const profile = await createCompanyProfile(req.session.token, { investor_company: req.body.investor_company })
     res.redirect(`/companies/${profile.investor_company.id}/investments/large-capital-profile`)
   } catch (err) {
-    res.locals.form = Object.assign({}, res.locals.form, {
-      errors: {
-        'summary': err.error.investor_company,
-      },
-    })
-    next()
+    if (err.statusCode === 400) {
+      set(res.locals, 'errors.summary', err.error.investor_company)
+      next()
+    } else {
+      next(err)
+    }
   }
 }
