+++ src/apps/investments/controllers/create/investor-profile.js
@@ -1,7 +1,7 @@
 const { get } = require('lodash')
 
 function renderCreateInvestorProfilePage (req, res) {
-  const errors = get(res.locals, 'form.errors')
+  const errors = get(res.locals, 'errors')
   return res
     .breadcrumb('Create a large capital investor profile')
     .render('investments/views/create/investor-profile', {
