+++ src/apps/investments/router-create-investor.js
@@ -1,5 +1,5 @@
-const { renderCreateInvestorProfilePage } = require('../controllers/create/investor-profile')
-const { createProfile } = require('../controllers/create/create-profile')
+const { renderCreateInvestorProfilePage } = require('./controllers/create/investor-profile')
+const { createProfile } = require('./controllers/create/create-profile')
 const router = require('express').Router()
 
 router
