+++ src/apps/investments/router-opportunities.js
@@ -1,0 +1,8 @@
+const setInvestmentTabItems = require('./middleware/investments-tab-items')
+const { renderOpportunitiesView } = require('./controllers/opportunities')
+
+const router = require('express').Router()
+
+router.get('/', setInvestmentTabItems, renderOpportunitiesView)
+
+module.exports = router
