--- src/apps/investments/routes/router-opportunities.js
@@ -1,8 +1,0 @@
-const setInvestmentTabItems = require('../middleware/investments-tab-items')
-const { renderOpportunitiesView } = require('../controllers/opportunities')
-
-const router = require('express').Router()
-
-router.get('/', setInvestmentTabItems, renderOpportunitiesView)
-
-module.exports = router
