--- src/apps/investments/routes/router-profiles.js
@@ -1,8 +1,0 @@
-const { renderProfilesView } = require('../controllers/profiles')
-const setInvestmentTabItems = require('../middleware/investments-tab-items')
-
-const router = require('express').Router()
-
-router.get('/', setInvestmentTabItems, renderProfilesView)
-
-module.exports = router
